# CRM/src/main/java/ru/arklual/crm/dto/TelegramSource.java
package ru.arklual.crm.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@AllArgsConstructor
public class TelegramSource {
    private String chatId;
}