# CRM/src/main/java/ru/arklual/crm/dto/responses/MessageResponse.java
package ru.arklual.crm.dto.responses;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class MessageResponse {

    String message;

}