# CRM/src/main/java/ru/arklual/crm/dto/responses/StatusResponse.java
package ru.arklual.crm.dto.responses;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class StatusResponse {

    String status;

}