# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost'
    description: Generated server url
paths:
...
  /api/auth/register:
    post:
      tags:
        - auth-controller
      summary: Регистрация нового пользователя
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Ошибка валидации запроса
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: Email уже зарегистрирован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/auth/login:
    post:
      tags:
        - auth-controller
      summary: Аутентификация пользователя
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        '200':
          description: Успешный вход
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Неверный формат запроса
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Неверный email или пароль
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/approval/reject:
    post:
      tags:
        - approval-controller
      summary: Отклонить пост
      operationId: rejectPost
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: postId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пост отклонён
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Недостаточно прав
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
  /api/approval/approve:
    post:
      tags:
        - approval-controller
      summary: Одобрить пост
      operationId: approvePost
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: postId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Пост одобрен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Недостаточно прав
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
  /api/approval/pending:
    get:
      tags:
        - approval-controller
      summary: Получить посты на модерации
      operationId: getPendingPosts
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Посты
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    AIFilter:
      type: object
      properties:
        trigger:
          type: string
        prompt:
          type: string
    AddPhotoFilter:
      type: object
      properties:
        trigger:
          type: string
        photoUrl:
          type: string
    Filter:
      type: object
      properties:
        addPhoto:
          $ref: '#/components/schemas/AddPhotoFilter'
        removePhoto:
          $ref: '#/components/schemas/RemovePhotoFilter'
        textReplace:
          $ref: '#/components/schemas/TextReplaceFilter'
        aiFilter:
          $ref: '#/components/schemas/AIFilter'
...
    RemovePhotoFilter:
      type: object
      properties:
        trigger:
          type: string
    Sink:
      type: object
      properties:
        telegramSink:
          $ref: '#/components/schemas/TelegramSink'
    Source:
      type: object
      properties:
        telegramSource:
          $ref: '#/components/schemas/TelegramSource'
    TelegramSink:
      type: object
      properties:
        chatId:
          type: string
    TelegramSource:
      type: object
      properties:
        chatId:
          type: string
    TextReplaceFilter:
      type: object
      properties:
        trigger:
          type: string
        pattern:
          type: string
        replacement:
          type: string
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
          maxLength: 2147483647
          minLength: 0
        password:
          type: string
      required:
        - email
        - name
        - password
    StatusResponse:
      type: object
      properties:
        status:
          type: string
    AuthRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
...
    Media:
      type: object
      properties:
        url:
          type: string
        mediaType:
          type: string
          enum:
            - PHOTO
            - VIDEO
            - AUDIO
            - UNRECOGNIZED
    PostResponse:
      type: object
      properties:
        id:
          type: string
        teamId:
          type: string
          format: uuid
        source:
          $ref: '#/components/schemas/Source'
        content:
          type: string
        media:
          $ref: '#/components/schemas/Media'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - DRAFT
            - PENDING_REVIEW
            - APPROVED
            - REJECTED
            - READY_TO_POST
            - POSTED
            - FAILED
            - UNRECOGNIZED
        requiresModeration:
          type: boolean
        sinks:
          type: array
          items:
            $ref: '#/components/schemas/Sink'
        filters:
          type: array
          items:
            $ref: '#/components/schemas/Filter'
        cleaned:
          type: boolean
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT